backend/pyproject.toml
[project]
name = "confirmation-parsers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "ollama>=0.5.2",
    "orjson>=3.11.1",
    "polars>=1.32.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
